[feat/#53] 카카오 로그인, 회원가입 컴포넌트 취합
